remove some duplicate code (i.e. req.logIn(...) by creating customSignup middleware function
