+++ routes/userRoutes.js
@@ -1,8 +1,7 @@
-const UsersDAO = require('../dao/usersDAO');
+const customSignup = require('../middleware/customSignup');
 const customAuth = require('../middleware/customPassportAuthenticate');
 const express = require('express');
 const router = express.Router();
-const bcrypt = require('bcryptjs');
 
 // Using the passport.authenticate middleware with our local strategy.
 router.post('/login', customAuth, function(req, res) {
@@ -10,23 +9,8 @@ router.post('/login', customAuth, function(req, res) {
 });
 
 // Route for signing up a user.
-router.post('/signup', async function(req, res, next) {
-  const hashedPswd = bcrypt.hashSync(req.body.password, 10);
-  const email = req.body.email;
-  try {
-    await UsersDAO.addUser(req.body.email, hashedPswd);
-    // Construct a new user object (FOR PASSPORT ONLY) containing just the email.
-    req.logIn({ email }, function(err) {
-      if (err) {
-        next(err);
-      } else {
-        res.json({ status: true, email, message: "You're now logged in!" });
-      }
-    });
-  } catch (e) {
-    console.error(`Error occurred while adding new user, ${e}`);
-    res.status(400).json({ status: false, message: 'Email already exists! Login instead.' });
-  }
+router.post('/signup', customSignup, customAuth, function(req, res, next) {
+  res.json({ status: true, email: req.user.email, message: "You're now logged in!" });
 });
 
 // Route for logging user out
